requirements.txt
streamlit>=1.36
python-dotenv>=1.0
openai>=1.40.0
pydantic>=2.8
